blogapp/requirements.txt
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
distro==1.8.0
dnspython==2.2.1
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
jsonschema==3.2.0
jwt==1.3.1
MarkupSafe==2.1.1
orjson==3.8.3
packaging==22.0
paramiko==2.12.0
passlib==1.7.4
psycopg-binary==3.1.6
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
PyNaCl==1.5.0
pyrsistent==0.19.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
texttable==1.6.7
typing_extensions==4.4.0
ujson==5.6.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websocket-client==0.59.0
websockets==10.4
